+++ src/containers/Form/flows/validation.js
@@ -1,0 +1,10 @@
+/*
+ * Copyright © 2015-2018 Serenova, LLC. All rights reserved.
+ */
+
+import { isEmpty } from 'serenova-js-utils/strings';
+
+export const formValidation = values => ({
+  name: isEmpty(values.get('name')) && 'Please enter a name',
+  type: !values.get('type') && 'Please select a type'
+});