+++ src/containers/Form/queues/__tests__/index.test.js
@@ -1,6 +1,3 @@
----
-to: src/containers/Form/<%= name %>/__tests__/index.test.js
----
 /*
  * Copyright © 2015-2018 Serenova, LLC. All rights reserved.
  */
@@ -9,15 +6,9 @@ import React from 'react';
 import { createStore } from 'redux';
 import { shallow } from 'enzyme';
 import { getCurrentForm } from '../../../../redux/modules/form/selectors';
-import <%= name %>Form, {
-  mapStateToProps
-} from '../';
-import {
-  getSelectedEntityId,
-  isInherited,
-  isCreating
-} from '../../../../redux/modules/entities/selectors';
-import { selectFormInitialValues,formSubmission, createFormName } from '../../../../redux/modules/form/selectors';
+import { mapStateToProps } from '../';
+import { getSelectedEntityId, isCreating } from '../../../../redux/modules/entities/selectors';
+import { selectFormInitialValues, formSubmission, createFormName } from '../../../../redux/modules/form/selectors';
 
 jest.mock('../../../../redux/modules/entities/selectors');
 jest.mock('../../../../redux/modules/form/selectors');
@@ -26,14 +17,10 @@ getSelectedEntityId.mockImplementation(() => 'mockId');
 isCreating.mockImplementation(() => true);
 selectFormInitialValues.mockImplementation(() => ({ active: true }));
 
-describe('<%= name %> Renders', () => {
+describe('queues Renders', () => {
   it('renders', () => {
     const store = createStore(state => state);
-    expect(
-      shallow(
-        <<%= name %>Form store={store}>Child</<%= name %>Form>
-      )
-    ).toMatchSnapshot();
+    expect(shallow(<queuesForm store={store}>Child</queuesForm>)).toMatchSnapshot();
   });
 });
 
@@ -50,14 +37,9 @@ describe('createFormName', () => {
 });
 
 describe('formSubmission', () => {
-  // Change values to match the form your making
-  // If you see this in a PR the unit tests may not have been completed
-  // const values = { name: 'mockName', value: 'mockValue' };
   const values = {
-    id: 'mockId',
     name: 'mockName',
-    description: 'mockDescription',
-    type: 'mockType'
+    description: 'mockDescription'
   };
   const dispatch = action => action;
   const props = { dirty: true };