+++ src/containers/Form/queues/__tests__/validation.test.js
@@ -1,0 +1,26 @@
+import { formValidation } from '../validation';
+import { Map } from 'immutable';
+
+describe('formValidation', () => {
+  it('returns proper object when required fields are provided', () => {
+    const values = new Map({
+      name: 'mockName',
+      description: 'mockDescription'
+    });
+    expect(formValidation(values)).toMatchSnapshot();
+  });
+  it('returns proper object (errors) when required fields are not provided', () => {
+    const values = new Map({
+      name: '',
+      description: ''
+    });
+    expect(formValidation(values)).toMatchSnapshot();
+  });
+  it('returns proper object (errors) when wrong values are provided', () => {
+    const values = new Map({
+      name: undefined,
+      description: undefined
+    });
+    expect(formValidation(values)).toMatchSnapshot();
+  });
+});