+++ src/containers/Form/queues/layout.js
@@ -1,0 +1,36 @@
+/*
+ * Copyright © 2015-2018 Serenova, LLC. All rights reserved.
+ */
+
+/**
+ *
+ * QueuesForm
+ *
+ */
+
+import React from 'react';
+import PropTypes from 'prop-types';
+import { DetailHeader, InputField } from 'cx-ui-components';
+
+export default function QueuesForm({ handleSubmit, isSaving, userHasUpdatePermission, key }) {
+  return (
+    <form onSubmit={handleSubmit} key={key}>
+      <DetailHeader text="Details" />
+      <InputField name="name" label="Name *" componentType="input" inputType="text" disabled={isSaving || !userHasUpdatePermission} />
+      <InputField
+        name="description"
+        label="Description"
+        componentType="textarea"
+        inputType="text"
+        disabled={isSaving || !userHasUpdatePermission}
+      />
+    </form>
+  );
+}
+
+QueuesForm.propTypes = {
+  key: PropTypes.string,
+  handleSubmit: PropTypes.func.isRequired,
+  isSaving: PropTypes.bool,
+  userHasUpdatePermission: PropTypes.bool
+};