+++ src/containers/SidePanelDetails/flows/index.js
@@ -1,0 +1,8 @@
+/*
+ * Copyright © 2015-2018 Serenova, LLC. All rights reserved.
+ */
+
+import { connect } from 'react-redux';
+import FlowsDetailsPanel from './layout';
+
+export default connect()(FlowsDetailsPanel);