+++ src/containers/SidePanelDetails/queues/__tests__/layout.test.js
@@ -1,0 +1,42 @@
+/*
+ * Copyright © 2015-2018 Serenova, LLC. All rights reserved.
+ */
+
+import React from 'react';
+import { shallow } from 'enzyme';
+
+import QueuesDetailsPanel from '../layout';
+
+describe('<queuesDetailsPanel />', () => {
+  let customItem;
+  beforeEach(() => {
+    customItem = {
+      name: 'Test',
+      description: 'Test'
+    };
+  });
+  it('renders queues detailsPanel', () => {
+    const rendered = shallow(
+      <QueuesDetailsPanel
+        id="0000-0000-0000-0000-0000"
+        className="details-panel"
+        userHasUpdatePermission={true}
+        children={'Mock Child'}
+        item={customItem}
+      />
+    );
+    expect(rendered).toMatchSnapshot();
+  });
+  it('renders queues detailsPanel with no update permision', () => {
+    const rendered = shallow(
+      <QueuesDetailsPanel
+        id="0000-0000-0000-0000-0000"
+        className="details-panel"
+        userHasUpdatePermission={false}
+        children={'Mock Child'}
+        item={customItem}
+      />
+    );
+    expect(rendered).toMatchSnapshot();
+  });
+});